chg: [ui] tooltip: compact properties, scrollable JSON, shared title fit
Make the tooltip match the sidebar — much tighter property rows, a scrollable container so a trailing JSON block is no longer clipped, and the auto-fit/identifier title treatment (extracted to titleFit.ts + an entity-title-name mixin shared with the main header).
