+++ src/ui/elements/Sidebar/MainHeader.ts
@@ -9,50 +9,7 @@ import { graphEdgeIcon, graphMultiSelectNode } from '../../icons'
 import type { EdgeSelection, NodeSelection } from '../../../interfaces/GraphInteractions'
 import { tryResolveHTMLElement } from '../../../utils/Getters'
 import { createNodePreview } from '../../../utils/NodePreview'
-import { createCopyButton } from './PropertyList'
-
-// Title auto-fit bounds: shrink a long title from MAX down to MIN px, wrapped
-// over at most MAX_LINES lines, before giving up and switching to the
-// type-aware fallback.
-const TITLE_MAX_PX = 16
-const TITLE_MIN_PX = 12
-const TITLE_LINE_HEIGHT = 1.3
-const TITLE_MAX_LINES = 2
-
-// A single hidden canvas reused to measure text width for middle-truncation.
-let textMeasurer: CanvasRenderingContext2D | null = null
-function measureTextWidth(text: string, font: string): number {
-    if (!textMeasurer) textMeasurer = document.createElement('canvas').getContext('2d')
-    if (!textMeasurer) return text.length * 8
-    textMeasurer.font = font
-    return textMeasurer.measureText(text).width
-}
-
-function elementFont(el: HTMLElement): string {
-    const s = getComputedStyle(el)
-    return `${s.fontWeight} ${s.fontSize} ${s.fontFamily}`
-}
-
-/** A title with no whitespace reads as an identifier (id, URL, hash, onion…). */
-function looksLikeIdentifier(text: string): boolean {
-    return !/\s/.test(text.trim())
-}
-
-/** Keep the head and tail of a too-long string, eliding the middle: `abcd…wxyz`. */
-function middleTruncate(text: string, availPx: number, font: string): string {
-    if (availPx <= 0 || measureTextWidth(text, font) <= availPx) return text
-    const ellipsis = '…'
-    let lo = 1, hi = text.length - 1, best = ellipsis
-    while (lo <= hi) {
-        const keep = (lo + hi) >> 1
-        const head = Math.ceil(keep / 2)
-        const tail = Math.floor(keep / 2)
-        const candidate = text.slice(0, head) + ellipsis + text.slice(text.length - tail)
-        if (measureTextWidth(candidate, font) <= availPx) { best = candidate; lo = keep + 1 }
-        else hi = keep - 1
-    }
-    return best
-}
+import { fitEntityTitle } from './titleFit'
 
 
 export class SidebarMainHeader extends UIComponent {
@@ -313,7 +270,7 @@ export class SidebarMainHeader extends UIComponent {
      * ends kept) plus a copy button, since middle-elision replaces the text.
      */
     private renderTitle(nameElem: HTMLElement, actionElem: HTMLElement | null, text: string): void {
-        this.fitCurrentTitle = () => this.fitTitle(nameElem, actionElem, text)
+        this.fitCurrentTitle = () => fitEntityTitle(nameElem, actionElem, text)
         this.titleLastWidth = -1
         requestAnimationFrame(() => this.refitTitle())
     }
@@ -326,38 +283,6 @@ export class SidebarMainHeader extends UIComponent {
         if (width === this.titleLastWidth) return
         this.titleLastWidth = width
         this.fitCurrentTitle()
-    }
-
-    private fitTitle(nameElem: HTMLElement, actionElem: HTMLElement | null, text: string): void {
-        // Reset to the auto-fit base state (normal wrap, no clamp, full text).
-        nameElem.className = 'pvt-mainheader-nodeinfo-name'
-        nameElem.style.fontSize = ''
-        nameElem.removeAttribute('title')
-        nameElem.textContent = text
-        actionElem?.replaceChildren()
-
-        const avail = nameElem.clientWidth
-        if (avail <= 0) return // collapsed / not yet laid out — the observer refits later
-
-        // 1) Auto-fit: the whole title, shrunk just enough to fit two lines.
-        for (let size = TITLE_MAX_PX; size >= TITLE_MIN_PX; size--) {
-            nameElem.style.fontSize = `${size}px`
-            if (nameElem.scrollHeight <= Math.ceil(size * TITLE_LINE_HEIGHT * TITLE_MAX_LINES) + 1) return
-        }
-
-        // 2) Too large even at the floor size → type-aware fallback.
-        nameElem.style.fontSize = ''
-        nameElem.title = text
-        if (looksLikeIdentifier(text)) {
-            nameElem.classList.add('is-identifier')
-            // Add the copy button first so it claims its cell, then middle-elide
-            // to the width that's actually left — otherwise the elided text is a
-            // touch too wide and the button spills past the sidebar edge.
-            actionElem?.appendChild(createCopyButton(text))
-            nameElem.textContent = middleTruncate(text, nameElem.clientWidth, elementFont(nameElem))
-        } else {
-            nameElem.classList.add('is-clamp')
-        }
     }
 
     /* Private methods */
