+++ src/ui/elements/Sidebar/mainHeader.scss
@@ -38,36 +38,7 @@
             font-family: 'Segoe UI', sans-serif;
 
             .pvt-mainheader-nodeinfo-name {
-                font-size: 16px;
-                font-weight: 600;
-                color: var(--pvt-text-color-6);
-                line-height: 1.3;
-                user-select: all;
-                // Auto-fit base: wrap normally, breaking a long token only when
-                // forced. JS shrinks the font (down to 12px) to fit two lines.
-                white-space: normal;
-                overflow-wrap: anywhere;
-                word-break: normal;
-
-                // Fallback A — prose too long to fit: clean two-line clamp.
-                &.is-clamp {
-                    display: -webkit-box;
-                    -webkit-line-clamp: 2;
-                    line-clamp: 2;
-                    -webkit-box-orient: vertical;
-                    overflow: hidden;
-                }
-
-                // Fallback B — identifier too long to fit: monospace, one line,
-                // middle-elided by JS (both ends kept); copy button carries the
-                // full value.
-                &.is-identifier {
-                    font-family: ui-monospace, 'SF Mono', Menlo, Consolas, monospace;
-                    font-size: 13.5px;
-                    letter-spacing: -0.01em;
-                    white-space: nowrap;
-                    overflow: hidden;
-                }
+                @include pvt.entity-title-name;
             }
 
             .pvt-mainheader-nodeinfo-subtitle {
